+++ packages/handlersjs-core/lib/public-api.ts
@@ -4,6 +4,8 @@ export * from './handlers/handler';
 export * from './handlers/passthrough-handler';
 export * from './handlers/sequence-handler';
 export * from './handlers/waterfall-handler';
+export * from './handlers/typed-pipe-through.handler';
+export * from './handlers/pipe-through.handler';
 export * from './logging/console-logger';
 export * from './logging/logger-level';
 export * from './logging/logger';