+++ packages/handlersjs-http/lib/models/http-handler-request.ts
@@ -1,6 +1,8 @@
+import { HttpMethod } from './http-method';
+
 export interface HttpHandlerRequest<B = any> {
   url: URL;
-  method: 'CONNECT' | 'DELETE' | 'GET' | 'HEAD' | 'OPTIONS' | 'PATCH' | 'POST' | 'PUT' | 'TRACE';
+  method: HttpMethod;
   parameters?: { [key: string]: string };
   headers: { [key: string]: string };
   body?: B;